Move pose implementation to right module
